Crate google_doubleclicksearch2

This documentation was generated from doubleclicksearch crate version 0.1.11+20160120, where 20160120 is the exact revision of the doubleclicksearch:v2 schema built by the mako code generator v0.1.11.

Everything else about the doubleclicksearch v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

conversion

get, insert, patch, update and update availability

reports

generate, get, get file and request

saved columns

list

Download supported by ...

get file reports

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.reports().generate(...).doit()
let r = hub.reports().get_file(...).doit()
let r = hub.reports().get(...).doit()
let r = hub.reports().request(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-doubleclicksearch2 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_doubleclicksearch2 as doubleclicksearch2;
use doubleclicksearch2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use doubleclicksearch2::Doubleclicksearch;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Doubleclicksearch::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.reports().get_file("reportId", -21)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Availability
A message containing availability data relevant to DoubleClick Search.

Conversion
A conversion containing data relevant to DoubleClick Search.

ConversionGetCall
Retrieves a list of conversions from a DoubleClick Search engine account.

ConversionInsertCall
Inserts a batch of new conversions into DoubleClick Search.

ConversionList
A list of conversions.

ConversionMethods
A builder providing access to all methods supported on conversion resources. It is not used directly, but through the Doubleclicksearch hub.

ConversionPatchCall
Updates a batch of conversions in DoubleClick Search. This method supports patch semantics.

ConversionUpdateAvailabilityCall
Updates the availabilities of a batch of floodlight activities in DoubleClick Search.

ConversionUpdateCall
Updates a batch of conversions in DoubleClick Search.

CustomDimension
A message containing the custome dimension.

CustomMetric
A message containing the custome metric.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Doubleclicksearch
Central instance to access all Doubleclicksearch related resource activities

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Report
A DoubleClick Search report. This object contains the report request, some report metadata such as currency code, and the generated report rows or report files.

ReportApiColumnSpec
A request object used to create a DoubleClick Search report.

ReportFiles
Asynchronous report only. Contains a list of generated report files once the report has succesfully completed.

ReportGenerateCall
Generates and returns a report immediately.

ReportGetCall
Polls for the status of a report request.

ReportGetFileCall
Downloads a report file encoded in UTF-8.

ReportMethods
A builder providing access to all methods supported on report resources. It is not used directly, but through the Doubleclicksearch hub.

ReportRequest
A request object used to create a DoubleClick Search report.

ReportRequestCall
Inserts a report request into the reporting system.

ReportRequestFilters
A list of filters to be applied to the report.

ReportRequestOrderBy
Synchronous report only. A list of columns and directions defining sorting to be performed on the report rows.

ReportRequestReportScope
The reportScope is a set of IDs that are used to determine which subset of entities will be returned in the report. The full lineage of IDs from the lowest scoped level desired up through agency is required.

ReportRequestTimeRange
If metrics are requested in a report, this argument will be used to restrict the metrics to a specific time range.

ReportRow
A row in a DoubleClick Search report.

SavedColumn
A saved column

SavedColumnList
A list of saved columns. Advertisers create saved columns to report on Floodlight activities, Google Analytics goals, or custom KPIs. To request reports with saved columns, you'll need the saved column names that are available from this list.

SavedColumnListCall
Retrieve the list of saved columns for a specified advertiser.

SavedColumnMethods
A builder providing access to all methods supported on savedColumn resources. It is not used directly, but through the Doubleclicksearch hub.

UpdateAvailabilityRequest
The request to update availability.

UpdateAvailabilityResponse
The response to a update availability request.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.