Module google_doubleclicksearch2::api

Structs

Availability
A message containing availability data relevant to DoubleClick Search.

Conversion
A conversion containing data relevant to DoubleClick Search.

ConversionGetCall
Retrieves a list of conversions from a DoubleClick Search engine account.

ConversionInsertCall
Inserts a batch of new conversions into DoubleClick Search.

ConversionList
A list of conversions.

ConversionMethods
A builder providing access to all methods supported on conversion resources. It is not used directly, but through the Doubleclicksearch hub.

ConversionUpdateAvailabilityCall
Updates the availabilities of a batch of floodlight activities in DoubleClick Search.

ConversionUpdateCall
Updates a batch of conversions in DoubleClick Search.

CustomDimension
A message containing the custom dimension.

CustomMetric
A message containing the custom metric.

Doubleclicksearch
Central instance to access all Doubleclicksearch related resource activities

Report
A DoubleClick Search report. This object contains the report request, some report metadata such as currency code, and the generated report rows or report files.

ReportApiColumnSpec
A request object used to create a DoubleClick Search report.

ReportFiles
Asynchronous report only. Contains a list of generated report files once the report has successfully completed.

ReportGenerateCall
Generates and returns a report immediately.

ReportGetCall
Polls for the status of a report request.

ReportGetFileCall
Downloads a report file encoded in UTF-8.

ReportMethods
A builder providing access to all methods supported on report resources. It is not used directly, but through the Doubleclicksearch hub.

ReportRequest
A request object used to create a DoubleClick Search report.

ReportRequestCall
Inserts a report request into the reporting system.

ReportRequestFilters
A list of filters to be applied to the report.\ The maximum number of filters per request is 300.

ReportRequestOrderBy
Synchronous report only. A list of columns and directions defining sorting to be performed on the report rows.\ The maximum number of orderings per request is 300.

ReportRequestReportScope
The reportScope is a set of IDs that are used to determine which subset of entities will be returned in the report. The full lineage of IDs from the lowest scoped level desired up through agency is required.

ReportRequestTimeRange
If metrics are requested in a report, this argument will be used to restrict the metrics to a specific time range.

ReportRow
A row in a DoubleClick Search report.

SavedColumn
A saved column

SavedColumnList
A list of saved columns. Advertisers create saved columns to report on Floodlight activities, Google Analytics goals, or custom KPIs. To request reports with saved columns, you'll need the saved column names that are available from this list.

SavedColumnListCall
Retrieve the list of saved columns for a specified advertiser.

SavedColumnMethods
A builder providing access to all methods supported on savedColumn resources. It is not used directly, but through the Doubleclicksearch hub.

UpdateAvailabilityRequest
The request to update availability.

UpdateAvailabilityResponse
The response to a update availability request.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.