Struct google_doubleclicksearch2::Conversion

pub struct Conversion {
    pub custom_dimension: Option<Vec<CustomDimension>>,
    pub ad_group_id: Option<String>,
    pub segmentation_id: Option<String>,
    pub attribution_model: Option<String>,
    pub campaign_id: Option<String>,
    pub revenue_micros: Option<String>,
    pub advertiser_id: Option<String>,
    pub quantity_millis: Option<String>,
    pub count_millis: Option<String>,
    pub criterion_id: Option<String>,
    pub conversion_timestamp: Option<String>,
    pub floodlight_order_id: Option<String>,
    pub segmentation_type: Option<String>,
    pub click_id: Option<String>,
    pub custom_metric: Option<Vec<CustomMetric>>,
    pub ds_conversion_id: Option<String>,
    pub engine_account_id: Option<String>,
    pub conversion_modified_timestamp: Option<String>,
    pub currency_code: Option<String>,
    pub segmentation_name: Option<String>,
    pub state: Option<String>,
    pub ad_id: Option<String>,
    pub agency_id: Option<String>,
    pub conversion_id: Option<String>,
    pub type_: Option<String>,
}

A conversion containing data relevant to DoubleClick Search.

This type is not used in any activity, and only used as part of another schema.

Fields

custom_dimension

Custom dimensions for the conversion, which can be used to filter data in a report.

ad_group_id

DS ad group ID.

segmentation_id

The numeric segmentation identifier (for example, DoubleClick Search Floodlight activity ID).

attribution_model

Attribution model name. This field is ignored.

campaign_id

DS campaign ID.

revenue_micros

The revenue amount of this TRANSACTION conversion, in micros.

advertiser_id

DS advertiser ID.

quantity_millis

The quantity of this conversion, in millis.

count_millis

The number of conversions, formatted in millis (conversions multiplied by 1000). This field is ignored.

criterion_id

DS criterion (keyword) ID.

conversion_timestamp

The time at which the conversion took place, in epoch millis UTC.

floodlight_order_id

The advertiser-provided order id for the conversion.

segmentation_type

The segmentation type of this conversion (for example, FLOODLIGHT).

click_id

DS click ID for the conversion.

custom_metric

Custom metrics for the conversion.

ds_conversion_id

DS conversion ID.

engine_account_id

DS engine account ID.

conversion_modified_timestamp

The time at which the conversion was last modified, in epoch millis UTC.

currency_code

The currency code for the conversion's revenue. Should be in ISO 4217 alphabetic (3-char) format.

segmentation_name

The friendly segmentation identifier (for example, DoubleClick Search Floodlight activity name).

state

The state of the conversion, that is, either ACTIVE or REMOVED. Note: state DELETED is deprecated.

ad_id

DS ad ID.

agency_id

DS agency ID.

conversion_id

Advertiser-provided ID for the conversion, also known as the order ID.

type_

The type of the conversion, that is, either ACTION or TRANSACTION. An ACTION conversion is an action by the user that has no monetarily quantifiable value, while a TRANSACTION conversion is an action that does have a monetarily quantifiable value. Examples are email list signups (ACTION) versus ecommerce purchases (TRANSACTION).