Struct google_doubleclicksearch2::ConversionMethods

pub struct ConversionMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on conversion resources. It is not used directly, but through the Doubleclicksearch hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_doubleclicksearch2 as doubleclicksearch2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use doubleclicksearch2::Doubleclicksearch;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Doubleclicksearch::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `patch(...)`, `update(...)` and `update_availability(...)`
// to build up your call.
let rb = hub.conversion();

Methods

impl<'a, C, A> ConversionMethods<'a, C, A>

fn insert(&self, request: ConversionList) -> ConversionInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a batch of new conversions into DoubleClick Search.

Arguments

request - No description provided.

fn get(&self, agency_id: &str, advertiser_id: &str, engine_account_id: &str, end_date: i32, row_count: i32, start_date: i32, start_row: u32) -> ConversionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of conversions from a DoubleClick Search engine account.

Arguments

agencyId - Numeric ID of the agency.

advertiserId - Numeric ID of the advertiser.

engineAccountId - Numeric ID of the engine account.

endDate - Last date (inclusive) on which to retrieve conversions. Format is yyyymmdd.

rowCount - The number of conversions to return per call.

startDate - First date (inclusive) on which to retrieve conversions. Format is yyyymmdd.

startRow - The 0-based starting index for retrieving conversions results.

fn update_availability(&self, request: UpdateAvailabilityRequest) -> ConversionUpdateAvailabilityCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the availabilities of a batch of floodlight activities in DoubleClick Search.

Arguments

request - No description provided.

fn patch(&self, request: ConversionList, advertiser_id: &str, agency_id: &str, end_date: i32, engine_account_id: &str, row_count: i32, start_date: i32, start_row: u32) -> ConversionPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a batch of conversions in DoubleClick Search. This method supports patch semantics.

Arguments

request - No description provided.

advertiserId - Numeric ID of the advertiser.

agencyId - Numeric ID of the agency.

endDate - Last date (inclusive) on which to retrieve conversions. Format is yyyymmdd.

engineAccountId - Numeric ID of the engine account.

rowCount - The number of conversions to return per call.

startDate - First date (inclusive) on which to retrieve conversions. Format is yyyymmdd.

startRow - The 0-based starting index for retrieving conversions results.

fn update(&self, request: ConversionList) -> ConversionUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a batch of conversions in DoubleClick Search.

Arguments

request - No description provided.