Struct google_doubleclicksearch2::ReportApiColumnSpec

pub struct ReportApiColumnSpec {
    pub group_by_column: Option<bool>,
    pub product_report_perspective: Option<String>,
    pub end_date: Option<String>,
    pub saved_column_name: Option<String>,
    pub custom_dimension_name: Option<String>,
    pub header_text: Option<String>,
    pub column_name: Option<String>,
    pub platform_source: Option<String>,
    pub start_date: Option<String>,
    pub custom_metric_name: Option<String>,
}

A request object used to create a DoubleClick Search report.

This type is not used in any activity, and only used as part of another schema.

Fields
group_by_column: Option<bool>

Synchronous report only. Set to true to group by this column. Defaults to false.

product_report_perspective: Option<String>

Returns metrics only for a specific type of product activity. Accepted values are: - "sold": returns metrics only for products that were sold - "advertised": returns metrics only for products that were advertised in a Shopping campaign, and that might or might not have been sold

end_date: Option<String>

Inclusive day in YYYY-MM-DD format. When provided, this overrides the overall time range of the report for this column only. Must be provided together with startDate.

saved_column_name: Option<String>

Name of a saved column to include in the report. The report must be scoped at advertiser or lower, and this saved column must already be created in the DoubleClick Search UI.

custom_dimension_name: Option<String>

Segments a report by a custom dimension. The report must be scoped to an advertiser or lower, and the custom dimension must already be set up in DoubleClick Search. The custom dimension name, which appears in DoubleClick Search, is case sensitive. If used in a conversion report, returns the value of the specified custom dimension for the given conversion, if set. This column does not segment the conversion report.

header_text: Option<String>

Text used to identify this column in the report output; defaults to columnName or savedColumnName when not specified. This can be used to prevent collisions between DoubleClick Search columns and saved columns with the same name.

column_name: Option<String>

Name of a DoubleClick Search column to include in the report.

platform_source: Option<String>

The platform that is used to provide data for the custom dimension. Acceptable values are "floodlight".

start_date: Option<String>

Inclusive date in YYYY-MM-DD format. When provided, this overrides the overall time range of the report for this column only. Must be provided together with endDate.

custom_metric_name: Option<String>

Name of a custom metric to include in the report. The report must be scoped to an advertiser or lower, and the custom metric must already be set up in DoubleClick Search. The custom metric name, which appears in DoubleClick Search, is case sensitive.