Struct google_doubleclicksearch2::ReportMethods

pub struct ReportMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on report resources. It is not used directly, but through the Doubleclicksearch hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_doubleclicksearch2 as doubleclicksearch2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use doubleclicksearch2::Doubleclicksearch;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Doubleclicksearch::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `generate(...)`, `get(...)`, `get_file(...)` and `request(...)`
// to build up your call.
let rb = hub.reports();

Methods

impl<'a, C, A> ReportMethods<'a, C, A>

fn get_file(&self, report_id: &str, report_fragment: i32) -> ReportGetFileCall<'a, C, A>

Create a builder to help you perform the following task:

Downloads a report file encoded in UTF-8.

Arguments

reportId - ID of the report.

reportFragment - The index of the report fragment to download.

fn generate(&self, request: ReportRequest) -> ReportGenerateCall<'a, C, A>

Create a builder to help you perform the following task:

Generates and returns a report immediately.

Arguments

request - No description provided.

fn get(&self, report_id: &str) -> ReportGetCall<'a, C, A>

Create a builder to help you perform the following task:

Polls for the status of a report request.

Arguments

reportId - ID of the report request being polled.

fn request(&self, request: ReportRequest) -> ReportRequestCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a report request into the reporting system.

Arguments

request - No description provided.