Struct google_doubleclicksearch2::ReportRequest

pub struct ReportRequest {
    pub order_by: Option<Vec<ReportRequestOrderBy>>,
    pub report_scope: Option<ReportRequestReportScope>,
    pub max_rows_per_file: Option<i32>,
    pub statistics_currency: Option<String>,
    pub time_range: Option<ReportRequestTimeRange>,
    pub start_row: Option<i32>,
    pub row_count: Option<i32>,
    pub report_type: Option<String>,
    pub columns: Option<Vec<ReportApiColumnSpec>>,
    pub filters: Option<Vec<ReportRequestFilters>>,
    pub include_removed_entities: Option<bool>,
    pub include_deleted_entities: Option<bool>,
    pub verify_single_time_zone: Option<bool>,
    pub download_format: Option<String>,
}

A request object used to create a DoubleClick Search report.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

generate reports (request)

request reports (request)

Fields
order_by: Option<Vec<ReportRequestOrderBy>>

Synchronous report only. A list of columns and directions defining sorting to be performed on the report rows.

report_scope: Option<ReportRequestReportScope>

The reportScope is a set of IDs that are used to determine which subset of entities will be returned in the report. The full lineage of IDs from the lowest scoped level desired up through agency is required.

max_rows_per_file: Option<i32>

Asynchronous report only. The maximum number of rows per report file. A large report is split into many files based on this field. Acceptable values are 1000000 to 100000000, inclusive.

statistics_currency: Option<String>

Specifies the currency in which monetary will be returned. Possible values are: usd, agency (valid if the report is scoped to agency or lower), advertiser (valid if the report is scoped to * advertiser or lower), or account (valid if the report is scoped to engine account or lower).

time_range: Option<ReportRequestTimeRange>

If metrics are requested in a report, this argument will be used to restrict the metrics to a specific time range.

start_row: Option<i32>

Synchronous report only. Zero-based index of the first row to return. Acceptable values are 0 to 50000, inclusive. Defaults to 0.

row_count: Option<i32>

Synchronous report only. The maxinum number of rows to return; additional rows are dropped. Acceptable values are 0 to 10000, inclusive. Defaults to 10000.

report_type: Option<String>

Determines the type of rows that are returned in the report. For example, if you specify reportType: keyword, each row in the report will contain data about a keyword. See the Types of Reports reference for the columns that are available for each type.

columns: Option<Vec<ReportApiColumnSpec>>

The columns to include in the report. This includes both DoubleClick Search columns and saved columns. For DoubleClick Search columns, only the columnName parameter is required. For saved columns only the savedColumnName parameter is required. Both columnName and savedColumnName cannot be set in the same stanza.

filters: Option<Vec<ReportRequestFilters>>

A list of filters to be applied to the report.

include_removed_entities: Option<bool>

Determines if removed entities should be included in the report. Defaults to false.

include_deleted_entities: Option<bool>

Determines if removed entities should be included in the report. Defaults to false. Deprecated, please use includeRemovedEntities instead.

verify_single_time_zone: Option<bool>

If true, the report would only be created if all the requested stat data are sourced from a single timezone. Defaults to false.

download_format: Option<String>

Format that the report should be returned in. Currently csv or tsv is supported.