Struct google_doubleclicksearch2::Conversion

pub struct Conversion {
    pub custom_dimension: Option<Vec<CustomDimension>>,
    pub ad_group_id: Option<String>,
    pub ds_conversion_id: Option<String>,
    pub campaign_id: Option<String>,
    pub revenue_micros: Option<String>,
    pub advertiser_id: Option<String>,
    pub count_millis: Option<String>,
    pub conversion_timestamp: Option<String>,
    pub click_id: Option<String>,
    pub attribution_model: Option<String>,
    pub currency_code: Option<String>,
    pub product_country: Option<i64>,
    pub state: Option<String>,
    pub device_type: Option<String>,
    pub ad_id: Option<String>,
    pub conversion_id: Option<String>,
    pub conversion_modified_timestamp: Option<String>,
    pub channel: Option<String>,
    pub product_id: Option<String>,
    pub segmentation_id: Option<String>,
    pub agency_id: Option<String>,
    pub store_id: Option<String>,
    pub quantity_millis: Option<String>,
    pub criterion_id: Option<String>,
    pub floodlight_order_id: Option<String>,
    pub segmentation_type: Option<String>,
    pub custom_metric: Option<Vec<CustomMetric>>,
    pub product_group_id: Option<String>,
    pub type_: Option<String>,
    pub segmentation_name: Option<String>,
    pub inventory_account_id: Option<String>,
    pub engine_account_id: Option<String>,
    pub product_language: Option<String>,
}

A conversion containing data relevant to DoubleClick Search.

This type is not used in any activity, and only used as part of another schema.

Fields

custom_dimension

Custom dimensions for the conversion, which can be used to filter data in a report.

ad_group_id

DS ad group ID.

ds_conversion_id

ID that DoubleClick Search generates for each conversion.

campaign_id

DS campaign ID.

revenue_micros

The revenue amount of this TRANSACTION conversion, in micros (value multiplied by 1000000, no decimal). For example, to specify a revenue value of "10" enter "10000000" (10 million) in your request.

advertiser_id

DS advertiser ID.

count_millis

Available to advertisers only after contacting DoubleClick Search customer support.

conversion_timestamp

The time at which the conversion took place, in epoch millis UTC.

click_id

DS click ID for the conversion.

attribution_model

Available to advertisers only after contacting DoubleClick Search customer support.

currency_code

The currency code for the conversion's revenue. Should be in ISO 4217 alphabetic (3-char) format.

product_country

The country registered for the Merchant Center feed that contains the product. Use an ISO 3166 code to specify a country.

state

The state of the conversion, that is, either ACTIVE or REMOVED. Note: state DELETED is deprecated.

device_type

The type of device on which the conversion occurred.

ad_id

DS ad ID.

conversion_id

For offline conversions, this is an ID that advertisers are required to provide. Advertisers can specify any ID that is meaningful to them. For online conversions, DS copies the dsConversionId or floodlightOrderId into this property depending on the advertiser's Floodlight instructions.

conversion_modified_timestamp

The time at which the conversion was last modified, in epoch millis UTC.

channel

Sales channel for the product. Acceptable values are: - "local": a physical store - "online": an online store

product_id

The product ID (SKU).

segmentation_id

The numeric segmentation identifier (for example, DoubleClick Search Floodlight activity ID).

agency_id

DS agency ID.

store_id

The ID of the local store for which the product was advertised. Applicable only when the channel is "local".

quantity_millis

The quantity of this conversion, in millis.

criterion_id

DS criterion (keyword) ID.

floodlight_order_id

The Floodlight order ID provided by the advertiser for the conversion.

segmentation_type

The segmentation type of this conversion (for example, FLOODLIGHT).

custom_metric

Custom metrics for the conversion.

product_group_id

DS product group ID.

type_

The type of the conversion, that is, either ACTION or TRANSACTION. An ACTION conversion is an action by the user that has no monetarily quantifiable value, while a TRANSACTION conversion is an action that does have a monetarily quantifiable value. Examples are email list signups (ACTION) versus ecommerce purchases (TRANSACTION).

segmentation_name

The friendly segmentation identifier (for example, DoubleClick Search Floodlight activity name).

inventory_account_id

ID that DS generates and uses to uniquely identify the inventory account that contains the product.

engine_account_id

DS engine account ID.

product_language

The language registered for the Merchant Center feed that contains the product. Use an ISO 639 code to specify a language.