Struct google_doubleclicksearch2::Report

pub struct Report {
    pub files: Option<Vec<ReportFiles>>,
    pub kind: Option<String>,
    pub rows: Option<Vec<ReportRow>>,
    pub request: Option<ReportRequest>,
    pub is_report_ready: Option<bool>,
    pub row_count: Option<i32>,
    pub statistics_time_zone: Option<String>,
    pub statistics_currency_code: Option<String>,
    pub id: Option<String>,
}

A DoubleClick Search report. This object contains the report request, some report metadata such as currency code, and the generated report rows or report files.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

generate reports (response)

get file reports (none)

get reports (response)

request reports (response)

Fields
files: Option<Vec<ReportFiles>>

Asynchronous report only. Contains a list of generated report files once the report has succesfully completed.

kind: Option<String>

Identifies this as a Report resource. Value: the fixed string doubleclicksearch#report.

rows: Option<Vec<ReportRow>>

Synchronous report only. Generated report rows.

request: Option<ReportRequest>

The request that created the report. Optional fields not specified in the original request are filled with default values.

is_report_ready: Option<bool>

Asynchronous report only. True if and only if the report has completed successfully and the report files are ready to be downloaded.

row_count: Option<i32>

The number of report rows generated by the report, not including headers.

statistics_time_zone: Option<String>

If all statistics of the report are sourced from the same time zone, this would be it. Otherwise the field is unset.

statistics_currency_code: Option<String>

The currency code of all monetary values produced in the report, including values that are set by users (e.g., keyword bid settings) and metrics (e.g., cost and revenue). The currency code of a report is determined by the statisticsCurrency field of the report request.

id: Option<String>

Asynchronous report only. Id of the report.